refactor: move block & object serialization to new useSerializer composable
- block deserialization might have functions reading/writing global data - variables, resources, etc. Helper methods cannot access codeStore but composable can. Hence grouping the serialization/deserialization functions dependent on data in useSerializer composable
