+++ frontend/src/stores/canvasStore.ts
@@ -1,7 +1,7 @@
 import { defineStore } from "pinia"
 import { ref, reactive, computed } from "vue"
 import type Block from "@/utils/block"
-import { getBlockCopy, getBlockInstance } from "@/utils/helpers"
+import { useSerializer } from "@/utils/useSerializer"
 
 import type StudioCanvas from "@/components/StudioCanvas.vue"
 import type { EditingMode, BlockOptions } from "@/types"
@@ -86,6 +86,8 @@ const useCanvasStore = defineStore("canvasStore", () => {
 	const showFragmentCanvas = computed(() => {
 		return editingMode.value === "fragment" || editingMode.value === "component" && fragmentData.value?.block
 	})
+
+	const { getBlockCopy, getBlockInstance } = useSerializer()
 
 	async function editOnCanvas(
 		block: Block,
