+++ frontend/src/stores/studioStore.ts
@@ -3,16 +3,12 @@ import router from "@/router/studio_router"
 import { defineStore } from "pinia"
 
 import {
-	getBlockInstance,
-	getRootBlock,
-	jsToJson,
-	getBlockCopyWithoutParent,
-	jsonToJs,
 	fetchApp,
 	fetchPage,
 	confirm,
 	getInitialVariableValue,
 } from "@/utils/helpers"
+import { useSerializer } from "@/utils/useSerializer"
 import { studioPages } from "@/data/studioPages"
 import { studioApps } from "@/data/studioApps"
 import { studioVariables } from "@/data/studioVariables"
@@ -150,6 +146,8 @@ const useStudioStore = defineStore("store", () => {
 	const selectedPage = ref<string | null>(null)
 	const savingPage = ref(false)
 	const settingPage = ref(false)
+
+	const { getBlockInstance, getRootBlock, getBlockCopyWithoutParent, jsToJson, jsonToJs } = useSerializer()
 
 	async function setPage(pageName: string) {
 		settingPage.value = true
