+++ frontend/src/utils/useCanvasHistory.ts
@@ -1,6 +1,7 @@
 import { ref, Ref } from "vue";
 import Block from "@/utils/block";
-import { generateId, getBlockInstance, getBlockString } from "@/utils/helpers";
+import { generateId } from "@/utils/helpers";
+import { useSerializer } from "@/utils/useSerializer"
 import { debounceFilter, pausableFilter, watchIgnorable } from "@vueuse/core";
 
 type CanvasState = {
@@ -65,6 +66,7 @@ export function useCanvasHistory(source: Ref<Block>, selectedBlockIds: Ref<Set<s
 	}
 
 	function createHistoryRecord() {
+		const { getBlockString } = useSerializer();
 		return {
 			block: getBlockString(source.value),
 			selectedBlockIds: selectedBlockIds.value,
@@ -72,6 +74,7 @@ export function useCanvasHistory(source: Ref<Block>, selectedBlockIds: Ref<Set<s
 	}
 
 	function setSource(value: CanvasState) {
+		const { getBlockInstance } = useSerializer();
 		ignorePrevAsyncBlockUpdates();
 		ignoreBlockUpdates(() => {
 			source.value = getBlockInstance(value.block);
